+++ transforms/move-pod-description-to-content/__testfixtures__/WithTypography.input.js
@@ -1,0 +1,9 @@
+import Pod from "carbon-react/lib/components/pod";
+import Typography from "carbon-react/lib/components/typography";
+
+const WithTypography = ({ children }) => {
+  return <>
+    <Pod description="Description" title="Title">{children}</Pod>
+    <Typography>accordion content</Typography>
+  </>;
+};